LayoutTests/fast/forms/time/time-stepup-stepdown.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Check stepUp() and stepDown() behavior for type=time.');

var input = document.createElement('input');

function setInputAttributes(min, max, step, value) {
    input.min = min;
    input.max = max;
    input.step = step;
    input.value = value;
}

function stepUp(value, step, max, optionalStepCount) {
    setInputAttributes(null, max, step, value);
    if (typeof optionalStepCount != "undefined")
        input.stepUp(optionalStepCount);
    else
        input.stepUp();
    return input.value;
}

function stepDown(value, step, min, optionalStepCount) {
    setInputAttributes(min, null, step, value);
    if (typeof optionalStepCount != "undefined")
        input.stepDown(optionalStepCount);
    else
        input.stepDown();
    return input.value;
}

input.type = 'time';
debug('Invalid value');
shouldBe('stepUp("", null, null)', '"00:01"');
shouldBe('stepDown("", null, null)', '"00:00"');
debug('Non-number arguments');
shouldBeEqualToString('stepUp("20:13", null, null, "0")', '20:13');
shouldBeEqualToString('stepDown("20:13", null, null, "0")', '20:13');
shouldBeEqualToString('stepUp("20:13", null, null, "foo")', '20:13');
shouldBeEqualToString('stepDown("20:13", null, null, "foo")', '20:13');
shouldBeEqualToString('stepUp("20:13", null, null, null)', '20:13');
shouldBeEqualToString('stepDown("20:13", null, null, null)', '20:13');
debug('Normal cases');
shouldBeEqualToString('stepUp("20:13", null, null)', '20:14');
shouldBeEqualToString('stepDown("20:13", null, null)', '20:12');
shouldBeEqualToString('stepUp("20:13", null, null, 10)', '20:23');
shouldBeEqualToString('stepDown("20:13", null, null, 11)', '20:02');
shouldBeEqualToString('stepUp("20:13", "4", null, 2)', '20:13:08');
shouldBeEqualToString('stepDown("20:13", "4", null, 3)', '20:12:48');
debug('Step=any');
shouldThrowErrorName('stepUp("20:13", "any", null)', "InvalidStateError");
shouldThrowErrorName('stepDown("20:13", "any", null)', "InvalidStateError");
debug('Overflow/underflow');
shouldBeEqualToString('stepUp("20:13", "3.40282346e+38", null)', '00:00:00');
shouldBeEqualToString('stepDown("20:13", "3.40282346e+38", null)', '00:00:00');
shouldBeEqualToString('stepUp("20:13", "1", "20:13")', '20:13:00');
shouldBeEqualToString('stepDown("20:13", "1", "20:13")', '20:13:00');
shouldBeEqualToString('stepUp("23:59", null, null)', '23:59');
shouldBeEqualToString('stepDown("00:00", null, null)', '00:00');

debug('');
</script>
</body>
</html>
